Try and fix ci.yml syntax
